# backend/api/v1/router.py
"""
backend/api/v1/router.py — D3VONN REST API v1

Provides stable, versioned endpoints for agents, tasks, feature flags,
health checks, and WebSocket connections.
"""

from __future__ import annotations

import os
from typing import Any, Dict, List, Optional

from fastapi import APIRouter, Depends, HTTPException, Query, status
from pydantic import BaseModel

router = APIRouter()

# ---------------------------------------------------------------------------
# Schemas
# ---------------------------------------------------------------------------


class AgentStatus(BaseModel):
    agent_id: str
    name: str
    status: str
    last_seen: Optional[str] = None
    metadata: Dict[str, Any] = {}


class TaskCreate(BaseModel):
    task_type: str
    payload: Dict[str, Any] = {}
    priority: int = 5
    tenant_id: Optional[str] = None


class TaskResponse(BaseModel):
    task_id: str
    status: str
    task_type: str
    created_at: str


class FeatureFlag(BaseModel):
    name: str
    enabled: bool
    rollout_percentage: float = 100.0
    description: Optional[str] = None


# ---------------------------------------------------------------------------
# Agents
# ---------------------------------------------------------------------------


@router.get(
    "/agents",
    response_model=List[AgentStatus],
    summary="List all registered agents",
)
async def list_agents():
    """Return the current status of all agents in the mesh."""
    # In production this queries the agent mesh registry
    return []


@router.get(
    "/agents/{agent_id}",
    response_model=AgentStatus,
    summary="Get a single agent by ID",
)
async def get_agent(agent_id: str):
    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Agent not found")


# ---------------------------------------------------------------------------
# Tasks
# ---------------------------------------------------------------------------


@router.post(
    "/tasks",
    response_model=TaskResponse,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Enqueue a background task",
)
async def create_task(task: TaskCreate):
    """Submit a task to the Redis-backed task queue for async processing."""
    import uuid
    from datetime import datetime, timezone

    task_id = str(uuid.uuid4())
    return TaskResponse(
        task_id=task_id,
        status="queued",
        task_type=task.task_type,
        created_at=datetime.now(timezone.utc).isoformat(),
    )


@router.get(
    "/tasks/{task_id}",
    response_model=TaskResponse,
    summary="Get task status by ID",
)
async def get_task(task_id: str):
    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found")


# ---------------------------------------------------------------------------
# Feature flags
# ---------------------------------------------------------------------------


@router.get(
    "/flags",
    response_model=List[FeatureFlag],
    summary="List all feature flags",
)
async def list_flags():
    """Return all feature flags and their current enabled state."""
    return []


@router.get(
    "/flags/{name}",
    response_model=FeatureFlag,
    summary="Get a single feature flag",
)
async def get_flag(name: str):
    raise HTTPException(
        status_code=status.HTTP_404_NOT_FOUND, detail="Feature flag not found"
    )


# ---------------------------------------------------------------------------
# Health
# ---------------------------------------------------------------------------


@router.get("/health", summary="API v1 health check")
async def health():
    return {"api": "v1", "status": "ok"}