# integrations/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Devonn.AI Platform API
  version: "1.0.0"
  description: AI orchestration and deployment platform API
  contact:
    name: Devonn.AI Platform Team
    url: https://github.com/wesship/supreme-ai-deployment-hub
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.d3vonn.io/v1
    description: Production
  - url: https://staging-api.d3vonn.io/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Local development

# Root-level security: all operations require bearer auth unless overridden
security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags: [System]
      security: []  # Health check is public
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Service unavailable

  /agents:
    get:
      summary: List all AI agents
      operationId: listAgents
      tags: [Agents]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a new AI agent
      operationId: createAgent
      tags: [Agents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /agents/{agentId}:
    get:
      summary: Get agent by ID
      operationId: getAgent
      tags: [Agents]
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /models:
    get:
      summary: List available AI models
      operationId: listModels
      tags: [Models]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /mesh/status:
    get:
      summary: Get agent mesh status
      operationId: getMeshStatus
      tags: [Mesh]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Mesh topology and health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Mesh unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthResponse:
      type: object
      required: [status, version, timestamp]
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        version:
          type: string
        timestamp:
          type: string
          format: date-time
        services:
          type: object
          additionalProperties:
            type: string

    Agent:
      type: object
      required: [id, name, model, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        created_at:
          type: string
          format: date-time
        config:
          type: object

    AgentList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    CreateAgentRequest:
      type: object
      required: [name, model]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        model:
          type: string
        config:
          type: object

    ModelList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              provider:
                type: string

    MeshStatus:
      type: object
      properties:
        nodes:
          type: integer
        edges:
          type: integer
        healthy_nodes:
          type: integer

    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object

  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"